Refactor folder structure for better organization
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/ATTSUMAN08/SwordofMagic7?shareId=XXXX-XXXX-XXXX-XXXX).
