--- src/main/java/swordofmagic7/Command/Player/AuctionCommand.java
@@ -1,40 +1,0 @@
-package swordofmagic7.Command.Player;
-
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import swordofmagic7.Auction;
-import swordofmagic7.Command.SomCommand;
-import swordofmagic7.Command.SomTabComplete;
-import swordofmagic7.Data.PlayerData;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class AuctionCommand implements SomCommand, SomTabComplete {
-    @Override
-    public boolean PlayerCommand(Player player, PlayerData playerData, String[] args) {
-        Auction.auctionCommand(playerData, args);
-        return true;
-    }
-
-    @Override
-    public boolean Command(CommandSender sender, String[] args) {
-        return false;
-    }
-
-    @Override
-    public List<String> PlayerTabComplete(Player player, PlayerData playerData, Command command, String[] args) {
-        return null;
-    }
-
-    @Override
-    public List<String> TabComplete(CommandSender sender, Command command, String[] args) {
-        List<String> complete = new ArrayList<>();
-        if (args.length == 1) {
-            complete.add("bet");
-            complete.add("sell");
-        }
-        return complete;
-    }
-}
